WordPressAuthenticator: Spliting SafariCredentialsService
